Version 1.3

+++ release/db.py
@@ -29,8 +29,10 @@ class dbapi:
         self.db.commit()
 
     def remove(self, eid):
-        self.db.execute('DELETE FROM events WHERE eid = ?', (eid,))
-        self.db.commit()
+        target_event_id = self.get_event_by_id(eid)
+        if target_event_id:
+            self.db.execute('DELETE FROM events WHERE eid = ?', (eid,))
+            self.db.commit()
 
     def get_events(self):
         cursor = self.db.execute('SELECT * FROM events ORDER BY date, time')
@@ -107,22 +109,34 @@ class dbapi:
 
     def check_and_add_next_routine(self):
         today = datetime.date.today()
+        tomorrow = today + datetime.timedelta(days=1)
+
+        # 如果当前时间已经过点了，那么下一个周期再添加这个事件。
+        def check_time(hour, minute):
+            now = datetime.datetime.now()
+            if now.hour > int(hour) or (now.hour == int(hour) and now.minute >= int(minute)):
+                return tomorrow
+            else:
+                return today
 
         routines = self.get_routines()
         for routine in routines:
             # 获取周期任务
             frequency, day_of_week, day_of_month, time, description = routine[1], routine[2], routine[3], routine[4], routine[5]
+
+            hour, minute = time.split(':')
+            today_or_tomorrow = check_time(hour, minute)
+
             if frequency == 'weekly':
-                next_date = self._get_next_weekly_date(today, day_of_week)
+                next_date = self._get_next_weekly_date(today_or_tomorrow, day_of_week)
             elif frequency == 'monthly':
-                next_date = self._get_next_monthly_date(today, day_of_month)
+                next_date = self._get_next_monthly_date(today_or_tomorrow, day_of_month)
 
             # 当前任务是否已经存在，如果不存在，添加任务
             check_cursor = self.db.execute('SELECT * FROM events WHERE date = ? AND time = ? AND event = ?', (next_date, time, description))
             check_list = [i for i in check_cursor]
-            print('cl:',check_list)
+
             if not check_list:
-                print('event added')
                 eid_set = self.get_all_eid()
                 eid = self.generate_new_eid(eid_set)
                 self.add(eid, description, next_date, time)
